+++ src/sentry/api/helpers/group_index/validators/inbox_details.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+
+class InboxDetailsValidator(serializers.Serializer):
+    # Support undo / snooze reasons
+    pass
