+++ src/sentry/api/helpers/group_index/validators/status_details.py
@@ -1,0 +1,58 @@
+from rest_framework import serializers
+
+from sentry.models import Release
+
+from . import InCommitValidator
+
+
+class StatusDetailsValidator(serializers.Serializer):
+    inNextRelease = serializers.BooleanField()
+    inRelease = serializers.CharField()
+    inCommit = InCommitValidator(required=False)
+    ignoreDuration = serializers.IntegerField()
+    ignoreCount = serializers.IntegerField()
+    # in minutes, max of one week
+    ignoreWindow = serializers.IntegerField(max_value=7 * 24 * 60)
+    ignoreUserCount = serializers.IntegerField()
+    # in minutes, max of one week
+    ignoreUserWindow = serializers.IntegerField(max_value=7 * 24 * 60)
+
+    def validate_inRelease(self, value):
+        project = self.context["project"]
+        if value == "latest":
+            try:
+                value = (
+                    Release.objects.filter(
+                        projects=project, organization_id=project.organization_id
+                    )
+                    .extra(select={"sort": "COALESCE(date_released, date_added)"})
+                    .order_by("-sort")[0]
+                )
+            except IndexError:
+                raise serializers.ValidationError(
+                    "No release data present in the system to form a basis for 'Next Release'"
+                )
+        else:
+            try:
+                value = Release.objects.get(
+                    projects=project, organization_id=project.organization_id, version=value
+                )
+            except Release.DoesNotExist:
+                raise serializers.ValidationError(
+                    "Unable to find a release with the given version."
+                )
+        return value
+
+    def validate_inNextRelease(self, value):
+        project = self.context["project"]
+        try:
+            value = (
+                Release.objects.filter(projects=project, organization_id=project.organization_id)
+                .extra(select={"sort": "COALESCE(date_released, date_added)"})
+                .order_by("-sort")[0]
+            )
+        except IndexError:
+            raise serializers.ValidationError(
+                "No release data present in the system to form a basis for 'Next Release'"
+            )
+        return value
